requirements.txt
letta-client>=0.1.0
pyyaml>=6.0.0
rich>=13.0.0

# Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0

